refactor plugin download
